Changed up all of the files, renamed folder to match what I want to call the project, figured out how to restructure the files and run build, hopefully this works who knows
