+++ .babelrc
@@ -1,0 +1,7 @@
+{
+    "presets": [
+        "@babel/react",
+        "@babel/env",
+        "@babel/preset-typescript"
+    ]
+}
