+++ .npmignore
@@ -1,0 +1,5 @@
+src
+examples
+.babelrc
+webpack.config.js
+public
