+++ examples/src/ReactWorldMap/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ReactWorldMap';
