+++ package.json
@@ -1,8 +1,9 @@
 {
-  "name": "react-world-map",
+  "name": "react-world-countries-map",
   "version": "1.0.0",
   "private": true,
-  "displayName": "React World Map",
+  "displayName": "React World Map of Countries",
+  "main": "./lib/index.tsx",
   "description": "A pure react component to visualize world maps with highlighted countries. Simple. Free. No Registration, No access token, No back-end service.",
   "authors": [
     "Yanivam (https://github.com/yanivam)"
@@ -18,21 +19,28 @@
     "@types/react": "^16.9.25",
     "@types/react-dom": "^16.9.5",
     "@types/react-map-gl": "^5.2.0",
+    "@types/webpack": "^4.41.8",
+    "babel-core": "^6.21.0",
     "d3-geo": "^1.11.9",
     "immutable": "^4.0.0-rc.12",
+    "path": "^0.12.7",
+    "prop-types": "^15.6.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-map-gl": "^5.2.3",
     "react-map-gl-draw": "^0.17.4-alpha.2",
     "react-scripts": "3.4.1",
     "react-svg-tooltip": "0.0.11",
-    "typescript": "~3.7.2"
+    "typescript": "~3.7.2",
+    "webpack": "^4.42.1"
+  },
+  "peerDependencies": {
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "build": "webpack --mode production"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -50,14 +58,25 @@
     ]
   },
   "devDependencies": {
-    "create-react-component-folder": "^0.2.0"
+    "@babel/core": "^7.9.0",
+    "@babel/loader": "^7.9.0",
+    "babel-loader": "^8.1.0",
+    "create-react-component-folder": "^0.2.0",
+    "gulp-babel": "^8.0.0",
+    "ts-loader": "^6.2.2",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
   },
   "keywords": [
     "react",
     "world map",
-    "typescript"
+    "map",
+    "component",
+    "typescript",
+    "svg",
+    "countries"
   ],
-  "bugs": "https://github.com/yanivam/react-world-map/issues",
-  "repository": "github:yanivam/react-world-map",
+  "bugs": "https://github.com/yanivam/react-world-countries-map/issues",
+  "repository": "github:yanivam/react-world-countries-map",
   "license": "Apache-2.0"
 }
